Fix potential race condition in running GC finalizers and I/O multiplexer.
We have no guarantee when a GC finalizer is run. Our finalizer used to tell
the I/O multiplexer that a socket is now unreachable and it would then be
properly unregistered from the multiplexer. However, because the finalizer
may be in another thread, the I/O multiplexer code and finalizer code may
use and modify the same data structure at the same time.

This change makes the I/O multiplexer use weak hash tables instead which
are presumed to be safe in a multithreaded environment. All current
finalizers are now trivial and deal with CFFI stuff.